Refactored all menu and draw functions
