+++ src/GravitySimulation_v2/GravitySimulation_v2/GravitySimulation_v2.vcxproj
@@ -149,6 +149,7 @@
     <ClCompile Include="simple_button.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="menu_utils.h" />
     <ClInclude Include="double_vector_ops.h" />
     <ClInclude Include="globals.h" />
     <ClInclude Include="grid.h" />
